+++ src/pythonwrench/functools.py
@@ -5,6 +5,8 @@ from typing import (
     Any,
     Callable,
     Generic,
+    Iterable,
+    Tuple,
     TypeVar,
     overload,
 )
@@ -13,6 +15,7 @@ from typing_extensions import ParamSpec
 
 from pythonwrench._core import _decorator_factory, return_none  # noqa: F401
 from pythonwrench.inspect import get_argnames
+from pythonwrench.typing import isinstance_generic
 
 T = TypeVar("T")
 P = ParamSpec("P")
@@ -24,6 +27,13 @@ class Compose(Generic[T, U]):
 
     @overload
     def __init__(self) -> None: ...
+
+    @overload
+    def __init__(
+        self,
+        fn0: Iterable[Callable[[T], T]],
+        /,
+    ) -> None: ...
 
     @overload
     def __init__(
@@ -73,7 +83,15 @@ class Compose(Generic[T, U]):
     @overload
     def __init__(self, *fns: Callable) -> None: ...
 
-    def __init__(self, *fns: Callable) -> None:
+    def __init__(self, *fns) -> None:
+        if isinstance_generic(fns, Tuple[Iterable[Callable]]):
+            fns = fns[0]
+        elif isinstance_generic(fns, Tuple[Callable, ...]):
+            pass
+        else:
+            msg = f"Invalid argument types {type(fns)=}."
+            raise TypeError(msg)
+
         super().__init__()
         self.fns = fns
 
@@ -124,3 +142,8 @@ def function_alias(alternative: Callable[P, U]) -> Callable[..., Callable[P, U]]
 def identity(x: T) -> T:
     """Identity function placeholder."""
     return x
+
+
+def repeat_fn(f: Callable[[T], T], n: int) -> Callable[[T], T]:
+    """Creates wrapper which call a function n items."""
+    return Compose([f] * n)